requirements.txt
streamlit
groq
PyPDF2
pdf2image
Pillow
faiss-cpu
numpy
sentence-transformers
